pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-customer-relationship"
version = "0.1.0"
description = "Customer Relationship Manager environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
  "openenv-core[core]>=0.2.2",
  "pydantic>=2.7.0",
  "openai>=1.40.0",
  "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]

[project.scripts]
server = "customer_relationship.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["customer_relationship", "customer_relationship.server"]
package-dir = { "customer_relationship" = ".", "customer_relationship.server" = "server" }
